+++ MangoAPI.Tests/DbContextFixture.cs
@@ -356,31 +356,31 @@ namespace MangoAPI.Tests
         {
             new ChatEntity
             {
-                CommunityType = CommunityType.PublicChannel,
+                CommunityType = (int) CommunityType.PublicChannel,
                 Title = "Extreme Code Main",
                 Id = SeedDataConstants.ExtremeCodeMainId,
             },
             new ChatEntity
             {
-                CommunityType = CommunityType.PrivateChannel,
+                CommunityType = (int) CommunityType.PrivateChannel,
                 Title = "Extreme Code C++",
                 Id = SeedDataConstants.ExtremeCodeCppId,
             },
             new ChatEntity
             {
-                CommunityType = CommunityType.ReadOnlyChannel,
+                CommunityType = (int) CommunityType.ReadOnlyChannel,
                 Title = "Extreme Code",
                 Id = SeedDataConstants.ExtremeCodeDotnetId,
             },
             new ChatEntity
             {
-                CommunityType = CommunityType.PrivateChannel,
+                CommunityType = (int) CommunityType.PrivateChannel,
                 Title = "Extreme Code Flood",
                 Id = SeedDataConstants.ExtremeCodeFloodId,
             },
             new ChatEntity
             {
-                CommunityType = CommunityType.DirectChat,
+                CommunityType = (int) CommunityType.DirectChat,
                 Title = "Petro / Szymon",
                 Id = SeedDataConstants.DirectPetroSzymon
             },
@@ -388,7 +388,7 @@ namespace MangoAPI.Tests
             {
                 Id = SeedDataConstants.WsbId,
                 Title = "WSB",
-                CommunityType = CommunityType.PublicChannel,
+                CommunityType = (int) CommunityType.PublicChannel,
                 Description = "WSB Public Group",
                 MembersCount = 5,
             }
@@ -400,187 +400,187 @@ namespace MangoAPI.Tests
             {
                 UserId = SeedDataConstants.PetroId,
                 ChatId = SeedDataConstants.WsbId,
-                RoleId = UserRole.Moderator,
+                RoleId = (int) UserRole.Moderator,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.SzymonId,
                 ChatId = SeedDataConstants.WsbId,
-                RoleId = UserRole.Owner,
+                RoleId = (int) UserRole.Owner,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.IlliaId,
                 ChatId = SeedDataConstants.WsbId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.ArslanbekId,
                 ChatId = SeedDataConstants.WsbId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.SerhiiId,
                 ChatId = SeedDataConstants.WsbId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KhachaturId,
                 ChatId = SeedDataConstants.ExtremeCodeMainId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.RazumovskyId,
                 ChatId = SeedDataConstants.ExtremeCodeMainId,
-                RoleId = UserRole.Admin,
+                RoleId = (int) UserRole.Admin,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KolbasatorId,
                 ChatId = SeedDataConstants.ExtremeCodeMainId,
-                RoleId = UserRole.Moderator,
+                RoleId = (int) UserRole.Moderator,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.AmelitId,
                 ChatId = SeedDataConstants.ExtremeCodeMainId,
-                RoleId = UserRole.Owner,
+                RoleId = (int) UserRole.Owner,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.RazumovskyId,
                 ChatId = SeedDataConstants.ExtremeCodeFloodId,
-                RoleId = UserRole.Owner,
+                RoleId = (int) UserRole.Owner,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KolbasatorId,
                 ChatId = SeedDataConstants.ExtremeCodeFloodId,
-                RoleId = UserRole.Moderator,
+                RoleId = (int) UserRole.Moderator,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KhachaturId,
                 ChatId = SeedDataConstants.ExtremeCodeFloodId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.AmelitId,
                 ChatId = SeedDataConstants.ExtremeCodeFloodId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.AmelitId,
                 ChatId = SeedDataConstants.ExtremeCodeCppId,
-                RoleId = UserRole.Owner,
+                RoleId = (int) UserRole.Owner,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.RazumovskyId,
                 ChatId = SeedDataConstants.ExtremeCodeCppId,
-                RoleId = UserRole.Admin,
+                RoleId = (int) UserRole.Admin,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KhachaturId,
                 ChatId = SeedDataConstants.ExtremeCodeCppId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KolbasatorId,
                 ChatId = SeedDataConstants.ExtremeCodeCppId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.RazumovskyId,
                 ChatId = SeedDataConstants.ExtremeCodeDotnetId,
-                RoleId = UserRole.Owner,
+                RoleId = (int) UserRole.Owner,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KolbasatorId,
                 ChatId = SeedDataConstants.ExtremeCodeDotnetId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KhachaturId,
                 ChatId = SeedDataConstants.ExtremeCodeDotnetId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.AmelitId,
                 ChatId = SeedDataConstants.ExtremeCodeDotnetId,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KhachaturId,
                 ChatId = SeedDataConstants.DirectKhachaturRazumovsky,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.RazumovskyId,
                 ChatId = SeedDataConstants.DirectKhachaturRazumovsky,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KolbasatorId,
                 ChatId = SeedDataConstants.DirectKolbasatorRazumovsky,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.RazumovskyId,
                 ChatId = SeedDataConstants.DirectKolbasatorRazumovsky,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.AmelitId,
                 ChatId = SeedDataConstants.DirectAmelitRazumovsky,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.RazumovskyId,
                 ChatId = SeedDataConstants.DirectAmelitRazumovsky,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KhachaturId,
                 ChatId = SeedDataConstants.DirectKhachaturKolbasator,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.KolbasatorId,
                 ChatId = SeedDataConstants.DirectKhachaturKolbasator,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.PetroId,
                 ChatId = SeedDataConstants.DirectPetroSzymon,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             },
             new UserChatEntity
             {
                 UserId = SeedDataConstants.SzymonId,
                 ChatId = SeedDataConstants.DirectPetroSzymon,
-                RoleId = UserRole.User,
+                RoleId = (int) UserRole.User,
             }
         };
